requirements.txt
# Core LangChain packages (modern versions)
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.1.0
langchain-openai>=0.1.0
langchain-ollama>=0.1.0

# Vector database
chromadb>=0.4.22

# Embeddings (via HuggingFace)
sentence-transformers>=2.5.2
huggingface_hub>=0.20.0

# OCR (Image → text)
pytesseract
pillow

# PDF loader
pypdf>=4.0.1

# Text loading & utilities
python-dotenv>=1.0.0
tiktoken>=0.5.2

# UI
streamlit>=1.29.0
